장비 현재 cpu 상태 조회 완료

+++ src/main/java/com/neonex/model/EqCpu.java
@@ -31,7 +31,7 @@ public class EqCpu {
     private String coreNum;
 
     @Column(name = "CPU_USAGE")
-    private int cpuUsage;
+    private double cpuUsage;
 }
